-Add handle the edit and delete HTTP requests if its not the postOwner

+++ SocialMedia-App/SocialMedia-App/Areas/User/Controllers/HomeController.cs
@@ -5,6 +5,7 @@ using SocialMedia.Models.Models;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity;
 using SocialMedia.Models.ViewModels;
+using Microsoft.EntityFrameworkCore;
 
 namespace SocialMedia_App.Areas.User.Controllers
 {
@@ -75,9 +76,16 @@ namespace SocialMedia_App.Areas.User.Controllers
 
         public IActionResult DeletePost(int id)
         {
+            var user = userManager.GetUserAsync(User).Result;
             Post searchedPost = db.Posts.Find(id);
+
+            if (searchedPost == null || searchedPost.PostOwnerId != user.Id)
+            {
+                return Unauthorized(); // Return 401 Unauthorized if the user is not the owner
+            }
+
             List<Comment> postComments = db.Comments.Where(c => c.PostId == id).ToList();
-            //Delete existing image
+            // Delete existing image
             if (searchedPost.ImageURL != null)
             {
                 string wwwRootPath = webHostEnvironment.WebRootPath;
@@ -89,6 +97,7 @@ namespace SocialMedia_App.Areas.User.Controllers
             db.SaveChanges();
             return RedirectToAction("Index");
         }
+
         public IActionResult DeleteComment(int id)
         {
             Comment searchedComment = db.Comments.Find(id);
@@ -99,28 +108,44 @@ namespace SocialMedia_App.Areas.User.Controllers
 
         public IActionResult EditPost(int id)
         {
+            var user = userManager.GetUserAsync(User).Result;
             PostViewModel viewModel = GetViewModel();
             viewModel.Post = db.Posts.Find(id);
+
+            if (viewModel.Post == null || viewModel.Post.PostOwnerId != user.Id)
+            {
+                return Unauthorized(); // Return 401 Unauthorized if the user is not the owner
+            }
+
             return View(viewModel.Post);
         }
+
         [HttpPost]
         public IActionResult EditPost(Post editedPost, IFormFile? file)
         {
+            var user = userManager.GetUserAsync(User).Result;
+            var originalPost = db.Posts.AsNoTracking().FirstOrDefault(p => p.PostId == editedPost.PostId);
+
+            if (originalPost == null || originalPost.PostOwnerId != user.Id)
+            {
+                return Unauthorized(); // Return 401 Unauthorized if the user is not the owner
+            }
+
             if (ModelState.IsValid)
             {
-                //Uploaded an new image
+                // Uploaded a new image
                 if (file != null)
                 {
                     string wwwRootPath = webHostEnvironment.WebRootPath;
                     string filename = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
                     string productPath = Path.Combine(wwwRootPath, @"images\posts");
-                    //Delete existing image
+                    // Delete existing image
                     if (editedPost.ImageURL != null)
                     {
                         string oldImagePath = Path.Combine(wwwRootPath, editedPost.ImageURL.TrimStart('\\'));
                         System.IO.File.Delete(oldImagePath);
                     }
-                    //Save new image
+                    // Save new image
                     using (FileStream s = new FileStream(Path.Combine(productPath, filename), FileMode.Create))
                     {
                         file.CopyTo(s);
@@ -130,14 +155,13 @@ namespace SocialMedia_App.Areas.User.Controllers
                 }
 
                 editedPost.DatePosted = DateTime.Now;
-                editedPost.PostOwnerId = signInManager.UserManager.GetUserAsync(User).Result.Id;
+                editedPost.PostOwnerId = user.Id;
                 db.Posts.Update(editedPost);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
             return View("EditPost", editedPost);
         }
-
         [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
         public IActionResult Error()
         {
